feat: add empty cart

+++ src/pages/basket/index.ts
@@ -1,5 +1,6 @@
 import './style.css';
 import { MainBasket } from './modules/main';
+import { cart } from '../start-page/modules/header';
 
 class Basket {
   main;
@@ -20,11 +21,22 @@ class Basket {
   }
 
   renderPage() {
-    this.main.renderItems();
-    this.main.renderSummary();
-    this.main.addItems();
-    this.main.deleteItems();
-    // this.main.promoCheck();
+    if (cart.currentCartProducts.length === 0) {
+      const emptyCart = document.createElement('h2');
+      emptyCart.innerText = 'CART IS EMPTY';
+      emptyCart.className = 'empty-cart';
+      const main = document.querySelector('.main');
+      if (main) {
+        main.innerHTML = '';
+        main.prepend(emptyCart);
+      }
+    } else {
+      this.main.renderItems();
+      this.main.renderSummary();
+      this.main.addItems();
+      this.main.deleteItems();
+      // this.main.promoCheck();
+    }
   }
 }
